requirements.txt
praw
matplotlib
python-dotenv